Added Project class with CRUD operations and linked it with Task class

+++ LayeredArchitecture/Monolith/TaskManager.Solution/TaskManager.BusinessLogic/Services/TaskService.cs
@@ -14,9 +14,9 @@ namespace TaskManager.Business.Services
             _taskRepo = taskRepo;
         }
 
-        public IEnumerable<TaskItem> GetAllTasks()
+        public IEnumerable<TaskItem> GetAllProjectTasks(int projectId)
         {
-            return _taskRepo.GetAll();
+            return _taskRepo.GetProjectTasks(projectId);
         }
 
         public TaskItem GetTaskById(int id)
@@ -26,26 +26,52 @@ namespace TaskManager.Business.Services
 
         public void AddTask(TaskItem task)
         {
+            if (string.IsNullOrWhiteSpace(task.Title))
+                throw new ArgumentException("Task title is required!");
+
+            if (task.Deadline < DateTime.Now)
+                throw new ArgumentException("Deadline can not be in the past!");
+
+            task.CreatedDate = DateTime.Now;
+            task.IsCompleted = false;
+
             _taskRepo.Add(task);
         }
 
-        public void UpdateTask(int id, TaskItem task)
+        public void UpdateTask(int id, TaskItem updatedTask)
         {
+            var task = _taskRepo.GetById(id);
+            if (task == null)
+                throw new ArgumentException("There is no task with that id");
+
+            if (string.IsNullOrWhiteSpace(updatedTask.Title))
+                throw new ArgumentException("Task title is required!");
+
+            if (updatedTask.Deadline < DateTime.Now)
+                throw new ArgumentException("Deadline can not be in the past!");
+
+            task.Title = updatedTask.Title;
+            task.Description = updatedTask.Description;
+            task.Deadline = updatedTask.Deadline;
             _taskRepo.Update(id, task);
         }
 
         public void DeleteTask(int id)
         {
+            var task = _taskRepo.GetById(id);
+            if (task == null)
+                throw new ArgumentException("There is no task with that id");
+
             _taskRepo.Delete(id);
         }
 
         public void CompleteTask(int id)
         {
-            /*var task = _taskRepo.GetById(id);
+            var task = _taskRepo.GetById(id);
             if (task == null) return;
 
             task.IsCompleted = true;
-            _taskRepo.Update(task);*/
+            _taskRepo.Update(id, task);
         }
     }
 }
